Progress on the slide controllers
We want some Python code to be executed in the template so we can have
code which needs to know about the template to make some decisions
(e.g., emit multiple slides for a TOC, even though there's only one
slide directive). The pagination depends on the layout, which is why
that code needs to be part of the template.