Example started of how I envision animations
This is a rough framework of how it should work. Think this is doable
but we need to make the SVG renderer more flexible to also accept
unnamed blobs. Or, better yet: We reference as hash keys only the
original source hash value (of the input SVG) and then add a key that
specifies which layers were visible.