Beginning of ExecRenderer
This is completely dysfunctional so far. But I think we have two issues:
(1) Identifying which command is executed from the args[0] list (so we
can hash it) and (2) I don't think the hash is honored even for the
image renderer, which would mean no forced re-rendering when the image
changes. Need to look into that.