Fixed the bug with double skipping slides
In presentation mode, the intersect event needs to be ignored, otherwise
that event in itself fires another goto_slide(), which causes advancing,
which might cause another event. Messy. Previously not noticed this
because it only happens on large monitors (full HD is insufficient to
get the advanced slides into the viewport, therefore the event never
fires).