Refactored features and slide information
We want to have slide information available to all slides of the
presentation, not generate individual CSS pieces for it. Therefore we
now attach it to the globalThis element once and reference it only
further down. This makes the feedback controller obsolete and the only
remaining function (adding the "info" feature) it would have has been
replaced by adding feature dependencies in the config file.