Implement deterministic session handling
We had opportunistic session handling before where the timer code would
simply listen to the first BroadcastMessage that came around and
accepted that as the "paired" session ID. This is stupid, but it is
*especially* stupid during development where I routinely have 3 or 4
presentation open at once and somehow the timer ALWAYS attaches to the
wrong window, leading me to chase ghost bugs.

Enough of that. Just pass the session ID to the window via the #hash and
be done with it.