Started work on spellchecker
Started work on spellchecker. It's not working just yet. We want to do
this efficiently, so we want to issuue only one call to LanguageTool,
but then we need to reverse back the offsets of text to chunks (which
are grouped in semantic groups). Quite a mess right now, but could work
soon.