Implemented spellchecker command line tool
This now is usable. It can emit simple vim syntax so that files can be
efficiently spellchecked. It's not very convenient yet: does not support
languages other than en_US, does not allow for adding to a dictionary,
does not have configuration options (such as specifying the JAR file of
LanguageTool only once and then having it as a default argument). But it
works and hence, fixes #1.