Add fulljson output to spellchecker, ignore whitespace rule
When we have an abbreviation as the last element in brackets (e.g.,
this is <s:ac>AES</s:ac>) the chunks that are generated are "...ths is
", "AES", ")". Mistakenly, LanguageTool then warns about a space before
the closing braces even though there are none (possibly because a new
chunk is always interpreted as a at least one space). Ignore the faulty
rule by default.

Added the fulljson output facility which eases debugging of such issues
tremendously.