Changed evim format and added calling vim directly
The syntax to format errormessages is rather obscure in vim and I could
not get it to work with having the evim data behind the %m message
specifier. So I now put the message in front, which messes up
readability but allows me to parse it. Since the vim invocation is
rather complex then, we call vim directly from pyradium if the user
wants that.