Beginning of style parameters implementation
We want to be able to pass style parameters, i.e., some form of
parameterizaton of the templates. One example would be light fonts vs.
heavy fonts that could just switch a CSS setting. For this the language
specifcation, similar to argparse, should be specified in the
configuration.json file of the template so that it can be nicely
displayed. Implementation of this has begun.