Fix the "file not found" error during saving in vim
Vim deletes the target file and moves in a replacement file. If during
that period, inotifywait triggers a rendering, it fails and the whole
loop collapses. This is very annoying, especially since the terminal in
the render loop is usually in the background and so the document isn't
updated while working on it. The soluton is in that case to retry
rendering after a short interval until it succeeds. Fixes #8.