Preparation to support external SVG animations
Having SVG animations outside pyradium could be a useful feature that I
would like to support, but right now it's deeply embedded in the
animation controller code. Decouple the two pieces of code so that we
can call them from an external script (e.g., to render multiple SVGs or
PDFs from individual frames). Right now the animation logic in the
controller is unchanged, but should in the future also use this SVG
animation core.