Added s:file hook
We want to have the ability to include arbitrary files, but also
obfuscate their filename so that if we include "foobar1.txt", people
cannot prematurely poke around searching for "foobar2.txt". We use an
extremely simple HMACish construction using MD5 for this. Good enough
for what we're trying to achieve.