Validate missing fonts in SVG files
Validation currently happens only on files (not literal data that is
included inside the XML) and parsing of the font-family style dictionary
is rudimentary (it may be allowed that multiple entries are possible,
e.g., style="Arial,'Nimbus Sans'" -- that we cannot parse currently).
Detection of SVG files is also made exclusively by the extension of the
filename (".svg") instead of honoring the "filetype" attribute or doing
proper MIME type guessing. That said, it should be good enough for the
overwhelming majority of use cases. Good enough for me. Fixes #46.