Significant improvement in variable substitution syntax
We're evaluating inside an environment in which we disallow access to
__builtins__. This means, however, that when inside our eval()
expression's f-string, we call datetime.datetime().strftime(), this
throws an exception because for some reason unclear to me, strftime() of
datetime.datetime calls __import__ (which is not present, since we
deliberately remove it from builtins). However, if we wrap
datetime.datetime in a proxy object, this works and leads to much nicer
syntax as well. So we're doing that.