Progress on scalable cursors
Scaling cursors is way more difficult than it seems. This is good
progress, we're parsing the SVG as XML, modify its DOM, then
re-serialize and push it into the CSS as verbatim data. However, the SVG
is still truncated and I'm not sure why (width and height seem to need
to scale with the square of scale_factor, which doesn't make sense to me
right now).