Refactoring of node replacment code
In some cases we need to cancel the recursive descent in some cases, as
shown in #56. However, this does not seem to work yet; this commit
refactors some of the code but ultimately does not fix the issue in #56
and also contains several debug statements that are printed on stdout.
WiP!