Seem to have fixed the root cause
The root cause was the usage of .replaceWholeText() and .wholeText on a
CharacterData DOM element; this does weird descending in itself, which
we never anticipated. It was just usage of the wrong method for the job
and caused the abort of recursive descent to fail, thereby accessing
wrong elements. Lots of debugging still left, will remove in a later
commit.