Fix scrolling when using "end" key
When pressing the "end" key, the last slide does not fully scroll into
the viewport because of a static offset computation (essentially, we are
scrolling up 100 pixels, the size of the top nav bar). With this patch,
we only apply that static shift when we are sure that it is required.
Fixes #67. Fixes #68.