Implemented MathJax support
Preliminary MathJax support added. Equation indentation is not
implemented yet.  I'm not sure this is a good idea, in particular
because it increases the size of every presentation by a 2 MiB blob of
code that is unconditionally used (even if the presentation does not
contain any equation). However, I do like the look. Undecided so far.