Shut linter up
The linter has the nit that subprocess.run should have an explicit
"check" kwarg passed. What a dumb and ridiculous linter rule.
subprocess.run already has a default value for that kwarg (check =
False) which is exactly the desired behavior. Let's not think about the
linter stupidity and just do as it wants us to.